修改了AR_开头的符号格式化操作，以前分平台需要%I64和%ll等问题，同一改成%qd，64bit整数或者%Id，根据当前平台选择提取整数范围

+++ trunk/Tools/GrammarDesigner/GrammarDesignerDoc.cpp
@@ -429,7 +429,7 @@ void CGrammarDesignerDoc::OnEditFindallreferences()
 						{
 								const ARSpace::lexToken_t *tok = &lx_set->token_set[i];
 
-								if(tok->count != sel.GetLength())
+								if(tok->count != (size_t)sel.GetLength())
 								{
 										continue;
 								}
